+++ extensions/bitwarden/src/components/RepromptForm.tsx
@@ -1,30 +1,15 @@
-import { Action, ActionPanel, Form, showToast, Toast } from "@raycast/api";
-import { Session } from "~/context/session";
+import { Action, ActionPanel, Form } from "@raycast/api";
 
 export type RepromptFormProps = {
-  session: Session;
   description: string;
-  onConfirm: () => void;
+  onConfirm: (password: string) => void;
 };
 
-/**
- * Form for confirming the master password.
- * This compares with the hashed master password.
- *
- * @param props.session The session instance.
- * @param props.description A description explaining why reprompting is required.
- * @param props.onConfirm Callback if confirmation is successful.
- */
 const RepromptForm = (props: RepromptFormProps) => {
-  const { session, description, onConfirm } = props;
+  const { description, onConfirm } = props;
 
   async function onSubmit(values: { password: string }) {
-    if (!(await session.confirmMasterPassword(values.password))) {
-      showToast(Toast.Style.Failure, "Confirmation failed.");
-      return;
-    }
-
-    onConfirm();
+    onConfirm(values.password);
   }
 
   return (
@@ -36,7 +21,7 @@ const RepromptForm = (props: RepromptFormProps) => {
         </ActionPanel>
       }
     >
-      <Form.Description text={description} />
+      <Form.Description title="Confirmation Required for" text={description} />
       <Form.PasswordField autoFocus id="password" title="Master Password" />
     </Form>
   );