+++ extensions/bitwarden/src/types/session.ts
@@ -1,10 +1,8 @@
 export interface SessionState {
   readonly token: string | undefined;
-
   readonly isLoading: boolean;
   readonly isLocked: boolean;
   readonly isAuthenticated: boolean;
-
-  readonly repromptHash: string | undefined;
-  readonly passwordEnteredDate: Date | undefined;
+  readonly passwordHash: string | undefined;
+  readonly lastActivityTime: Date | undefined;
 }