+++ extensions/chatgpt/src/hooks/useAutoTTS.tsx
@@ -1,0 +1,7 @@
+import { getPreferenceValues } from "@raycast/api";
+
+export function useAutoTTS(): boolean {
+  return getPreferenceValues<{
+    isAutoTTS: boolean;
+  }>().isAutoTTS;
+}