+++ extensions/ifanr/src/type.d.ts
@@ -1,0 +1,5 @@
+export interface IListItem {
+  title: string;
+  url: string;
+  author: string;
+}