+++ extensions/vercast/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Fixed persistency of the selected team
 
+## [2.3] - 2023-03-28
+
+- Added gray dot icon for queued deployments
+
 ## [2.2] - 2022-09-08
 
 - Added a new command: `Search Components` for quickly searching and accessing the Vercel design library